requirements.txt
keyboard==0.13.5
protobuf==5.27.0
pyttsx3==2.90
RealtimeSTT==0.1.16
google-generativeai==0.6.0